Update index.html
In the calculation section:

The variables used in the calculations are incorrect. You should use the variables you defined earlier, but you are using the input field IDs instead. For example, instead of puntajeFundamentosInput, it should be puntajeFundamentos.
The variables numeroDeEvidencias, puntajeFinal, puntajeEvaluacionTecnica, and others are not defined. You need to use the correct variable names that you defined earlier.
In the display section:

When you try to display the results using console.log, you are using the input field IDs directly instead of using the variables that hold the input values. For example, instead of nombreInput, it should be nombre.
Similarly, you need to use the correct variable names for displaying other values such as puntajeEvaluacionTecnica, numeroDeEvidencias, puntajeIngles, totalTutorias, and puntajeFinal.